refactor: move initial helpers to utils
