+++ generators/api/expected/bar/bazQux/bazQuxAPI.ts
@@ -1,5 +1,5 @@
-import resolveFetch from 'helpers/resolveFetch'
 import BazQux from 'models/BazQux'
+import resolveFetch from 'utils/resolveFetch'
 
 export function fetchBazQux() {
 	return resolveFetch<BazQux>(fetch('/api/bazQux.json'))
