+++ generators/api/templates/fooAPI.ts
@@ -1,5 +1,5 @@
-import resolveFetch from 'helpers/resolveFetch'
 import <%= model %> from 'models/<%= model %>'
+import resolveFetch from 'utils/resolveFetch'
 
 export function fetch<%= model %>() {
 	return resolveFetch<<%= model %>>(fetch('/api/<%= camelCase(model) %>.json'))
