+++ generators/app/templates/src/utils/bootstrap.tsx
