+++ generators/app/templates/src/utils/createReducer.ts
