+++ generators/app/templates/src/utils/intl.ts
