+++ generators/app/templates/src/utils/resolveFetch.test.ts
