+++ generators/app/templates/src/utils/resolveFetch.ts
